refactor: 핵심 dto가 생성 역할을 가지도록 수정
